scripts/requirements.txt
azure-cosmos==4.7.0
slack-sdk==3.33.1